Rename entity to instance, Rentity to Entity
